refactor(runner): migrate chronicle.ts off setAtPath; delete the legacy write helpers
Follow-on to #3704. With v2-transaction's hot write path now on
`applyMutablePathWrite()`, the only remaining caller of
`Attestation.write` (and transitively `setAtPath`, `sparseArrayCopy`)
was Chronicle's working-copy management for commit-time conflict
detection. This PR migrates Chronicle onto the same mutate-in-place
helper, then deletes the legacy structurally-sharing implementations.

Mechanical changes:

- New module `packages/runner/src/storage/transaction/mutable-path-write.ts`
  holding `applyMutablePathWrite`, `applyArrayLengthWrite`, `MutableWriteResult`,
  `isContainerValue`, and `getValueTypeName`. Moved out of
  `v2-transaction.ts` so both v2-transaction and chronicle can import
  without a backwards dependency.
- `chronicle.ts` grows a thin `applyWriteToAttestation()` wrapper that
  maps Chronicle's `IAttestation`-shaped inputs to
  `applyMutablePathWrite`'s `FabricValue`-rooted form. The three
  internal call sites (Novelty.claim, Changes.applyWrite, Changes.put)
  route through it instead of `Attestation.write`.
- `attestation.ts` loses `setAtPath`, `sparseArrayCopy`, and the
  exported `write` function. ~190 lines deleted.

Semantic change to be aware of: Chronicle now auto-creates missing
intermediate containers on a nested write (matching what
v2-transaction has done since #3704), where the old setAtPath returned
a `NotFoundError`. The `workingCopy.value === undefined && path.length > 0`
case is still pre-checked inside `Chronicle.write` and still returns
NotFound, so the only behavior change is for "existing doc, missing
intermediate" -- a case not exercised by any test.

`docs/specs/sparse-array-preservation.md` is updated: the
"Attestation" section is replaced by a "Chronicle" section pointing at
the unified `applyMutablePathWrite` path, and the test-coverage
listing is repointed at `cell-core.test.ts` (which exercises the new
path end-to-end with sparse-array writes).

Tests:

- 468 runner / 120 memory / 35 data-model tests pass.
- The two direct Chronicle tests (`memory-v2-envelope-delete.test.ts`,
  `transaction-notfound.test.ts`) continue to pass with the new
  internal write helper.
- `frozen-reads-cache.bench.ts` and `compact-real.bench.ts` numbers are
  consistent with #3704's post-merge baseline (no regression from the
  extraction).
