+++ packages/runner/src/storage/transaction/attestation.ts
@@ -3,7 +3,6 @@ import { isRecord } from "@commonfabric/utils/types";
 import {
   type FabricObject,
   type FabricValue,
-  isArrayIndexPropertyName,
 } from "@commonfabric/data-model/fabric-value";
 import { toCompactDebugString } from "@commonfabric/data-model/value-debug";
 import type {
@@ -24,15 +23,6 @@ import { unclaimed } from "@commonfabric/memory/fact";
 import { getLogger } from "@commonfabric/utils/logger";
 import { LRUCache } from "@commonfabric/utils/cache";
 import { toTransactionDocumentValue } from "../v2-document.ts";
-
-/** Copies an array preserving sparse holes (unlike [...arr] which fills them with undefined). */
-const sparseArrayCopy = <T>(arr: T[]): T[] => {
-  const copy = new Array<T>(arr.length);
-  arr.forEach((v, i) => {
-    copy[i] = v;
-  });
-  return copy;
-};
 
 const logger = getLogger("attestation", {
   enabled: false,
@@ -73,203 +63,6 @@ export const UnsupportedMediaTypeError = (
     return this;
   },
 });
-
-/**
- * Sets a value at path using structural sharing. Only clones along the path.
- * Returns original reference if nothing changed (noop propagation).
- *
- * CT-1123 Phase 2: This replaces the O(N) JSON.parse(JSON.stringify()) deep clone
- * with O(D) structural sharing where D is path depth.
- *
- * NOTE: As of #3704, this is no longer on the v2-transaction write path
- * (which uses `applyMutablePathWrite` for in-place mutation). It is still
- * used by `chronicle.ts` for its working-copy management; once that
- * migrates, this and `write` below can be removed.
- */
-const setAtPath = (
-  root: FabricValue,
-  path: readonly MemoryAddressPathComponent[],
-  value: FabricValue,
-): { ok: FabricValue } | {
-  error: { at: number; type: string; notFound?: boolean };
-} => {
-  // Base case: empty path = replace root
-  if (path.length === 0) {
-    return { ok: value };
-  }
-
-  const [key, ...rest] = path;
-
-  // Type check: can't traverse through non-objects
-  if (root === null || root === undefined || typeof root !== "object") {
-    // Distinguish between undefined (path not found) vs primitive (type mismatch)
-    if (root === undefined) {
-      // Return at: -1 to indicate the error is at the parent level (the key that
-      // led here doesn't exist). After +1 adjustments during unwinding, this
-      // produces a path that includes the missing key, matching read semantics.
-      return { error: { at: -1, type: "undefined", notFound: true } };
-    }
-    const actualType = root === null ? "null" : typeof root;
-    return { error: { at: 0, type: actualType } };
-  }
-
-  // Handle arrays
-  if (Array.isArray(root)) {
-    // Special: array.length property
-    if (key === "length" && rest.length === 0) {
-      const newLen = value as number;
-      if (root.length === newLen) return { ok: root }; // noop
-      // Use slice for truncation, negative values, and non-finite values (NaN, Infinity)
-      // slice handles these edge cases correctly and matches previous behavior
-      if (newLen < root.length || newLen < 0 || !Number.isFinite(newLen)) {
-        return { ok: root.slice(0, newLen) };
-      } else {
-        // Extension: create array with new length (sparse slots preserved)
-        const extended = sparseArrayCopy(root);
-        extended.length = newLen;
-        return { ok: extended };
-      }
-    }
-
-    // Validate array key
-    if (!isArrayIndexPropertyName(key)) {
-      return { error: { at: 0, type: "array" } };
-    }
-
-    const index = Number(key);
-
-    // Terminal case
-    if (rest.length === 0) {
-      if (root[index] === value) return { ok: root }; // noop
-      const newArray = sparseArrayCopy(root);
-      if (value === undefined) {
-        delete newArray[index]; // creates hole, not splice
-      } else {
-        newArray[index] = value;
-      }
-      return { ok: newArray };
-    }
-
-    // Recursive case
-    const nested = root[index];
-    const result = setAtPath(nested, rest, value);
-    if ("error" in result) {
-      return {
-        error: {
-          at: result.error.at + 1,
-          type: result.error.type,
-          notFound: result.error.notFound,
-        },
-      };
-    }
-    if (result.ok === nested) return { ok: root }; // noop propagation
-    const newArray = sparseArrayCopy(root);
-    newArray[index] = result.ok as FabricValue;
-    return { ok: newArray };
-  }
-
-  // Handle objects
-  const obj = root as FabricObject;
-
-  // Terminal case
-  if (rest.length === 0) {
-    if (obj[key] === value) return { ok: root }; // noop
-    if (value === undefined) {
-      if (!(key in obj)) return { ok: root }; // delete non-existent = noop
-      const { [key]: _, ...without } = obj;
-      return { ok: without as FabricValue };
-    }
-    return { ok: { ...obj, [key]: value } };
-  }
-
-  // Recursive case
-  const nested = obj[key];
-  const result = setAtPath(nested, rest, value);
-  if ("error" in result) {
-    return {
-      error: {
-        at: result.error.at + 1,
-        type: result.error.type,
-        notFound: result.error.notFound,
-      },
-    };
-  }
-  if (result.ok === nested) return { ok: root }; // noop propagation
-  return { ok: { ...obj, [key]: result.ok as FabricValue } };
-};
-
-/**
- * Takes `source` attestation, `address` and `value` and produces derived
- * attestation with `value` set to a property that given `address` leads to
- * in the `source`. Fails with type mismatch error if provided `address` leads
- * to a non-object target, or NotFound error if the document doesn't exist.
- *
- * CT-1123 Phase 2: Now uses structural sharing via setAtPath() instead of
- * JSON.parse(JSON.stringify()) deep clone. Only clones objects along the
- * modified path, leaving siblings shared.
- *
- * NOTE: As of #3704, the v2-transaction write path no longer routes through
- * here -- it uses `applyMutablePathWrite` for in-place mutation. This
- * function is still used by `chronicle.ts` for its working-copy management.
- */
-export const write = (
-  source: IAttestation,
-  address: IMemoryAddress,
-  value: FabricValue,
-): Result<
-  IAttestation,
-  IStorageTransactionInconsistent | INotFoundError | ITypeMismatchError
-> => {
-  // Calculate relative path from source
-  const relativePath = address.path.slice(source.address.path.length);
-
-  // Root write: path lengths equal (empty relative path)
-  if (relativePath.length === 0) {
-    if (source.value === value) return { ok: source }; // noop
-    return { ok: { ...source, value } };
-  }
-
-  // Can't write nested path on undefined document
-  if (source.value === undefined) {
-    return {
-      error: NotFound(source, address, source.address.path),
-    };
-  }
-
-  // Apply structural sharing via setAtPath
-  const result = setAtPath(source.value, relativePath, value);
-
-  if ("error" in result) {
-    // Map error position to full address path
-    // result.error.at is the depth where the error occurred; +1 to include the failed key
-    const errorPath = address.path.slice(
-      0,
-      source.address.path.length + result.error.at + 1,
-    );
-
-    // Distinguish between NotFound (path doesn't exist) and TypeMismatch (wrong type)
-    if (result.error.notFound) {
-      // errorPath includes the missing key, matching read error semantics
-      return {
-        error: NotFound(source, address, errorPath),
-      };
-    }
-    return {
-      error: TypeMismatchError(
-        { ...address, path: errorPath },
-        result.error.type,
-        "write",
-      ),
-    };
-  }
-
-  // Noop: setAtPath returns original reference if nothing changed
-  if (result.ok === source.value) {
-    return { ok: source };
-  }
-
-  return { ok: { ...source, value: result.ok } };
-};
 
 /**
  * Reads requested `address` from the provided `source` attestation and either
